sre2_worker_agent

Module sre2_agent

here we will have node sre2_agent centered this will hold the node logic using prompts specific and structured output schema specific to any mini-node we have broken down the agents task into smaller ones easier to digest and compatible with local llm limitations so ‘conditional edges’ will be in the bigger function that coordinated the nodes

Structs

Sre2AgentHandler

Functions

run_commit

then it will be able to commit work

run_read

The sre agent will pass through different steps read the target file, write the target file with changes from task instruction, commit work, create a report on work done. thisis the read file: after having receive instruction on what work to do it will pick the right file and read its content for context awarness

run_report

finally it will be creating a report so that next agent get a nice overview of what has been done

run_write

after reading it will be able to write the content to the same file following new requirements

sre2_agent_node_work_orchestration